requirements.txt
discord.py
yt-dlp
spotipy
python-dotenv
PyNaCl
ffmpeg